requirements.txt
asgiref==3.8.1
gunicorn
Django==5.0.6
django-admin-rangefilter==0.13.0
django-import-export
MarkupPy==1.14
numpy
odfpy==1.4.1
openpyxl
psycopg==3.1.19
psycopg-binary==3.1.19
setuptools==69.5.1
simplejson==3.19.2
wheel==0.43.0
xlrd
xlwt==1.3.0
